+++ src/helpers/getConfigurationFile.ts
@@ -1,5 +1,5 @@
-import * as fs from 'fs';
 import * as path from 'path';
+import { readFile } from './readFile';
 import { ConfigurationFile } from '../types';
 import { getWorkspaceFolder } from './getWorkspaceFolder';
 
@@ -8,9 +8,8 @@ export function getConfigurationFile(): ConfigurationFile | null {
     const workspaceFolderPath = getWorkspaceFolder();
 
     if (workspaceFolderPath) {
-      const file = fs.readFileSync(
+      const file = readFile(
         path.resolve(workspaceFolderPath, 'codebot.config.json'),
-        'utf-8',
       );
 
       const configurationFile = JSON.parse(file);